+++ src/update.cpp
@@ -102,13 +102,13 @@ int graf( int age, int p0, int p1, int p2, int p3, int p4, int p5, int p6 )
    if( age < 15 )
       return p0;  // < 15
    else if( age <= 29 )
-      return ( int )( p1 + ( ( ( age - 15 ) * ( p2 - p1 ) ) / 15 ) );   /* 15..29 */
+      return ( p1 + ( ( ( age - 15 ) * ( p2 - p1 ) ) / 15 ) );   /* 15..29 */
    else if( age <= 44 )
-      return ( int )( p2 + ( ( ( age - 30 ) * ( p3 - p2 ) ) / 15 ) );   /* 30..44 */
+      return ( p2 + ( ( ( age - 30 ) * ( p3 - p2 ) ) / 15 ) );   /* 30..44 */
    else if( age <= 59 )
-      return ( int )( p3 + ( ( ( age - 45 ) * ( p4 - p3 ) ) / 15 ) );   /* 45..59 */
+      return ( p3 + ( ( ( age - 45 ) * ( p4 - p3 ) ) / 15 ) );   /* 45..59 */
    else if( age <= 79 )
-      return ( int )( p4 + ( ( ( age - 60 ) * ( p5 - p4 ) ) / 20 ) );   /* 60..79 */
+      return ( p4 + ( ( ( age - 60 ) * ( p5 - p4 ) ) / 20 ) );   /* 60..79 */
    else
       return p6;  // >= 80
 }
